+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - New #65: Add `IP_PATTERN` and `IP_REGEXP` constants  to `IpHelper` for checking IP of both IPv4 and IPv6 versions             
   (@arogachev)
+- New #65 Add `NEGATION_CHARACTER` constant to `IpRanges` used to negate ranges (@arogachev)
 - New #65 Add `isIpv4()`, `isIpv6()`, `isIp()` methods to `IpHelper` (@arogachev)
 
 ## 1.1.0 August 06, 2024